Trying stuff out.

+++ tests/toy_test.py
@@ -12,41 +12,32 @@ import matplotlib.pyplot as plt
 from tf_unet import image_gen
 from tf_unet import unet
 from tf_unet import util
-from utils.functions import binarize
+
+EPOCHS = 20
+TRAIN = False
 
 np.random.seed(98765)
 
-def train():
-    nx = 572
-    ny = 572
-
-    generator = image_gen.GrayScaleDataProvider(nx, ny, cnt = 20)
-    net = unet.Unet(
-        channels = generator.channels,
-        n_class = generator.n_class,
-        layers = 3,
-        features_root = 16
-    )
+def train_network(net, generator):
     trainer = unet.Trainer(
         net,
         optimizer = 'momentum',
         opt_kwargs = dict(momentum = 0.2)
     )
-    path = trainer.train(
+    trainer.train(
         generator,
         './unet_trained_toy',
         prediction_path = './prediction_toy',
-        training_iters = 32,
-        epochs = 1,
+        #training_iters = 32,
+        epochs = EPOCHS,
         display_step = 2
     )
-
-    return net, generator
     
 def predict(net, generator):
     x, y = generator(1)
     prediction = net.predict('./unet_trained_toy/model.ckpt', x)
     mask = prediction#[0, ..., 1] > 0.9
+    mask = np.select([prediction <= 0.5, prediction > 0.5], [np.zeros_like(prediction), np.ones_like(prediction)])
 
     print(np.shape(x))
     print(np.shape(y))
@@ -61,11 +52,11 @@ def predict(net, generator):
     print(np.amin(mask))
     print(np.amax(mask))
 
-    fig, ax = plt.subplots(1, 4, sharex = True, sharey = True, figsize = (12, 5))
+    fig, ax = plt.subplots(1, 4, figsize = (12, 4))
     ax[0].imshow(x[0, ..., 0], aspect = 'auto', cmap = plt.cm.gray)
     ax[1].imshow(y[0, ..., 1], aspect = 'auto', cmap = plt.cm.gray)
-    ax[2].imshow(prediction[0, ..., 1], aspect = 'auto', cmap = plt.cm.binary)
-    ax[3].imshow(mask[0, ..., 1], aspect = 'auto', cmap = plt.cm.binary)
+    ax[2].imshow(prediction[0, ..., 1], aspect = 'auto', cmap = plt.cm.gray)
+    ax[3].imshow(mask[0, ..., 1], aspect = 'auto', cmap = plt.cm.gray)
     ax[0].set_title('Input')
     ax[1].set_title('Ground truth')
     ax[2].set_title('Prediction')
@@ -74,17 +65,29 @@ def predict(net, generator):
     plt.draw()
     plt.show()
 
-def test():
-    net, generator = train()
+def test(train = True):
+    nx = 572
+    ny = 572
+    generator = image_gen.GrayScaleDataProvider(nx, ny, cnt = 20)
+    net = unet.Unet(
+        channels = generator.channels,
+        n_class = generator.n_class,
+        layers = 3,
+        features_root = 16
+    )
+
+    if train:
+        train_network(net, generator)
     predict(net, generator)
 
-def init():
-    shutil.rmtree('./unet_trained_toy', ignore_errors = True)
-    shutil.rmtree('./prediction_toy', ignore_errors = True)
+def init(remove_previous_net = True):
+    if remove_previous_net:
+        shutil.rmtree('./unet_trained_toy', ignore_errors = True)
+        shutil.rmtree('./prediction_toy', ignore_errors = True)
 
 def main():
-    init()
-    test()
+    init(remove_previous_net = TRAIN)
+    test(TRAIN)
 
 if __name__ == '__main__':
     main()
